refactor: 💡 Parser - Syntax Analysis
Refactoring of the current state of the code done
