+++ Makefile
@@ -5,7 +5,7 @@ NAME=minishell.a
 EXEC_NAME=minishell
 INC_DIR=include
 SRC_DIR=src
-SRC_FILES=main.c
+SRC_FILES=main.c parser_token_build.c parser_token_utils.c parser_token_debug.c 
 LIBFT_DIR=libft
 LIBFT=$(LIBFT_DIR)/libft.a
 SRCS=$(addprefix $(SRC_DIR)/, $(SRC_FILES))
