+++ include/minishell.h
@@ -12,8 +12,9 @@
 
 #include <stdio.h>
 #include <readline/readline.h>
+#include "../libft/include/libft.h"
 
-# define TERMINAL_PROMPT "🚀 $: "
+#define TERMINAL_PROMPT "🚀 $: "
 
 typedef enum e_token_type
 {
@@ -31,3 +32,10 @@ typedef struct s_token
 	t_token_type	type;
 	struct s_token	*next;
 }	t_token;
+
+void	print_tokens(t_token *tokens_head);
+int		token_lst_build(t_token **tokens_head, char *value);
+void	token_lst_clear(t_token **token_lst_head);
+void	token_lst_add_back(t_token **token_lst_head, t_token *token);
+/* DEGUB */
+char	*print_token_type(t_token_type type);
